License agreement and screenshot support for add-ons

+++ source/Playnite/ViewModels/AddonsViewModelBase_Online.cs
@@ -2,6 +2,7 @@
 using Playnite.Common.Web;
 using Playnite.Plugins;
 using Playnite.SDK;
+using Playnite.Windows;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -56,6 +57,51 @@ namespace Playnite.ViewModels
             this.resources = resources;
         }
 
+        public bool CheckAddonLicense(AddonManifest addon)
+        {
+            try
+            {
+                if (addon.UserAgreement != null)
+                {
+                    var acceptState = ExtensionInstaller.GetAddonLicenseAgreed(addon.AddonId);
+                    if (acceptState == null || acceptState < addon.UserAgreement.Updated)
+                    {
+                        var license = HttpDownloader.DownloadString(addon.UserAgreement.AgreementUrl);
+                        var licenseAgree = new LicenseAgreementViewModel(
+                            new LicenseAgreementWindowFactory(),
+                            license,
+                            addon.Name);
+
+                        if (licenseAgree.OpenView() == true)
+                        {
+                            ExtensionInstaller.AgreeAddonLicense(addon.AddonId);
+                            return true;
+                        }
+                        else
+                        {
+                            ExtensionInstaller.RemoveAddonLicenseAgreement(addon.AddonId);
+                            return false;
+                        }
+                    }
+                    else
+                    {
+                        return true;
+                    }
+                }
+                else
+                {
+                    return true;
+                }
+            }
+            catch (Exception e) when (!PlayniteEnvironment.ThrowAllErrors)
+            {
+                logger.Error(e, $"Failed to process addon license.");
+                dialogs.ShowErrorMessage(LOC.AddonErrorDownloadFailed, string.Empty);
+            }
+
+            return false;
+        }
+
         public void UpdateAddons()
         {
             var addons = UpdateAddonList.Where(a => a.Selected == true);
@@ -75,6 +121,13 @@ namespace Playnite.ViewModels
                         prg.CurrentProgressValue++;
                         prg.Text = string.Format(resources.GetString(LOC.AddonDownloadingAddon), update.Item.Name);
                         Thread.Sleep(500);
+
+                        if (!CheckAddonLicense(update.Item))
+                        {
+                            update.Status = AddonUpdateStatus.LicenseRejected;
+                            continue;
+                        }
+
                         var locaPath = update.Item.GetTargetDownloadPath();
                         FileSystem.DeleteFile(locaPath);
